Introduce WPCF7_Stripe_API class (not trait)
#465
