+++ modules/stripe/service.php
@@ -5,7 +5,6 @@ if ( ! class_exists( 'WPCF7_Service' ) ) {
 }
 
 class WPCF7_Stripe extends WPCF7_Service {
-	use WPCF7_Stripe_API;
 
 	private static $instance;
 	private $api_keys;
@@ -43,6 +42,14 @@ class WPCF7_Stripe extends WPCF7_Service {
 	}
 
 
+	public function api() {
+		if ( $this->is_active() ) {
+			$api = new WPCF7_Stripe_API( $this->api_keys['secret'] );
+			return $api;
+		}
+	}
+
+
 	public function get_api_keys() {
 		return $this->api_keys;
 	}
@@ -62,11 +69,6 @@ class WPCF7_Stripe extends WPCF7_Service {
 			'https://stripe.com/',
 			'stripe.com'
 		);
-	}
-
-
-	protected function log( $url, $request, $response ) {
-		wpcf7_log_remote_request( $url, $request, $response );
 	}
 
 
@@ -245,118 +247,4 @@ class WPCF7_Stripe extends WPCF7_Service {
 </form>
 <?php
 	}
-}
-
-
-/**
- * Trait for the Stripe API.
- *
- * @link https://stripe.com/docs/api
- */
-trait WPCF7_Stripe_API {
-
-	/**
-	 * Creates a Payment Intent.
-	 *
-	 * @link https://stripe.com/docs/api/payment_intents/create
-	 *
-	 * @param string|array $args Optional. Arguments to control behavior.
-	 * @return array|bool An associative array if 200 OK, false otherwise.
-	 */
-	public function create_payment_intent( $args = '' ) {
-		$args = wp_parse_args( $args, array(
-			'amount' => 0,
-			'currency' => '',
-			'receipt_email' => '',
-		) );
-
-		$endpoint = 'https://api.stripe.com/v1/payment_intents';
-
-		$request = array(
-			'headers' => $this->default_headers(),
-			'body' => $args,
-		);
-
-		$response = wp_remote_post( esc_url_raw( $endpoint ), $request );
-
-		if ( 200 != wp_remote_retrieve_response_code( $response ) ) {
-			if ( WP_DEBUG ) {
-				$this->log( $endpoint, $request, $response );
-			}
-
-			return false;
-		}
-
-		$response_body = wp_remote_retrieve_body( $response );
-		$response_body = json_decode( $response_body, true );
-
-		return $response_body;
-	}
-
-
-	/**
-	 * Retrieve a Payment Intent.
-	 *
-	 * @link https://stripe.com/docs/api/payment_intents/retrieve
-	 *
-	 * @param string $id Payment Intent identifier.
-	 * @return array|bool An associative array if 200 OK, false otherwise.
-	 */
-	public function retrieve_payment_intent( $id ) {
-		$endpoint = sprintf(
-			'https://api.stripe.com/v1/payment_intents/%s',
-			urlencode( $id )
-		);
-
-		$request = array(
-			'headers' => $this->default_headers(),
-		);
-
-		$response = wp_remote_get( esc_url_raw( $endpoint ), $request );
-
-		if ( 200 != wp_remote_retrieve_response_code( $response ) ) {
-			if ( WP_DEBUG ) {
-				$this->log( $endpoint, $request, $response );
-			}
-
-			return false;
-		}
-
-		$response_body = wp_remote_retrieve_body( $response );
-		$response_body = json_decode( $response_body, true );
-
-		return $response_body;
-	}
-
-
-	/**
-	 * Returns default set of HTTP request headers used for Stripe API.
-	 *
-	 * @link https://stripe.com/docs/building-plugins#setappinfo
-	 *
-	 * @return array An associative array of headers.
-	 */
-	private function default_headers() {
-		$app_info = array(
-			'name' => 'WordPress Contact Form 7',
-			'partner_id' => 'pp_partner_HHbvqLh1AaO7Am',
-			'url' => 'https://contactform7.com/',
-			'version' => WPCF7_VERSION,
-		);
-
-		$ua = array(
-			'lang' => 'php',
-			'lang_version' => PHP_VERSION,
-			'application' => $app_info,
-		);
-
-		$headers = array(
-			'Authorization' => sprintf( 'Bearer %s', $this->api_keys['secret'] ),
-			'Stripe-Version' => '2020-08-27',
-			'X-Stripe-Client-User-Agent' => json_encode( $ua ),
-		);
-
-		return $headers;
-	}
-
 }
